+++ foxes/models/wake_models/induction/rankine_half_body.py
@@ -260,14 +260,18 @@ class RankineHalfBody(WakeModel):
 
         """
 
-        # calc ambient wind vector
+        # calc ambient wind vector:
         ws0 = amb_results[FV.WS]
-        wind_vec = wd2uv(amb_results[FV.WD], ws0)
+        nx = wd2uv(amb_results[FV.WD])
+        wind_vec = nx * ws0[:, :, None]
 
-        # add ambient result to wake deltas
-        delta_uv = np.stack((wake_deltas["U"], wake_deltas["V"]), axis=2)
+        # wake deltas are in wake frame, rotate back to global frame:
+        ny = np.stack((-nx[:, :, 1], nx[:, :, 0]), axis=2)
+        delta_uv = wake_deltas["U"][:, :, None] * nx + wake_deltas["V"][:, :, None] * ny
+
+        # add ambient result to wake deltas:
         wind_vec += delta_uv
-        del delta_uv, ws0
+        del delta_uv, ws0, nx, ny
 
         # deduce WS and WD deltas:
         new_wd = uv2wd(wind_vec)