+++ docs/source/history.md
@@ -441,5 +441,7 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 
 ## v0.6.1
 
+- Bug fixes:
+  - Bug fixed for `RankineHalfBody` turbine induction model that produced wrong results for wind directions unequal 270 degrees
 
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v0.6.1](https://github.com/FraunhoferIWES/foxes/commits/v0.6.1)