+++ docs/source/history.md
@@ -441,6 +441,12 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 
 ## v0.6.1
 
+- Input:
+  - New ambient states: `TabStates`, single tab-file input
+- Data
+  - New static data: `winds100.tab`, an example tab file
+- Examples:
+  - New example: `tab_file`, demonstrating the usage of the `TabStates`
 - Bug fixes:
   - Bug fixed for `RankineHalfBody` turbine induction model that produced wrong results for wind directions unequal 270 degrees